working project
got a working .jar now
