+++ .gitignore
@@ -24,3 +24,8 @@ hs_err_pid*
 replay_pid*
 /ProjectGUI/nbproject/private/
 /ProjectGUI/build/
+/CharacterSheet/target/
+/ProjectGUI/dist/
+/CharacterProject/build/
+/CharacterProject/nbproject/private/
+/CharacterProject/dist/
