+++ CharacterProject/build.xml
@@ -7,11 +7,11 @@
 <!-- the Compile on Save feature is turned off for the project. -->
 <!-- You can turn off the Compile on Save (or Deploy on Save) setting -->
 <!-- in the project's Project Properties dialog box.-->
-<project name="ProjectGUI" default="default" basedir=".">
-    <description>Builds, tests, and runs the project ProjectGUI.</description>
+<project name="CharacterProject" default="default" basedir=".">
+    <description>Builds, tests, and runs the project CharacterProject.</description>
     <import file="nbproject/build-impl.xml"/>
+    <!--
 
-<!--
     There exist several targets which are by default empty and which can be 
     used for execution of your tasks. These targets are usually executed 
     before and after some main targets. They are: 
@@ -58,7 +58,7 @@
 
     An example of overriding the target for project execution could look like this:
 
-        <target name="run" depends="ProjectGUI-impl.jar">
+        <target name="run" depends="CharacterProject-impl.jar">
             <exec dir="bin" executable="launcher.exe">
                 <arg file="${dist.jar}"/>
             </exec>
