+++ CharacterProject/src/Model/BasicStatManager.java
