+++ CharacterProject/src/View/EditInventory.java
@@ -164,8 +164,6 @@ public class EditInventory extends javax.swing.JFrame {
         }
         //</editor-fold>
         //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
