+++ CharacterProject/src/View/EditSkills.java
@@ -665,8 +665,6 @@ public class EditSkills extends javax.swing.JFrame {
         }
         //</editor-fold>
         //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
