+++ CharacterProject/src/View/ErrorScreen.java
@@ -126,7 +126,6 @@ public class ErrorScreen extends javax.swing.JFrame {
             java.util.logging.Logger.getLogger(ErrorScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
         //</editor-fold>
-        //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
