+++ CharacterProject/src/View/MainScreen.java
@@ -902,8 +902,6 @@ public class MainScreen extends javax.swing.JFrame {
         }
         //</editor-fold>
         //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
