+++ CharacterProject/src/View/WarningScreenNew.java
@@ -5,23 +5,17 @@
 package View;
 
 import Controller.MenuManager;
-import Json.jsonData;
-import java.io.File;
-import javax.swing.JFileChooser;
-import javax.swing.filechooser.FileNameExtensionFilter;
 
 /**
  *
  * @author nsara
  */
-public class WarningScreenLoad extends javax.swing.JFrame {
+public class WarningScreenNew extends javax.swing.JFrame {
 
-    
-    private final jsonData dataCall = new jsonData();
     /**
      * Creates new form WarningScreen
      */
-    public WarningScreenLoad() {
+    public WarningScreenNew() {
         initComponents();
     }
 
@@ -103,17 +97,8 @@ public class WarningScreenLoad extends javax.swing.JFrame {
     }//GEN-LAST:event_jButton2ActionPerformed
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
+        MenuManager.newCharS();
         this.dispose();
-        JFileChooser file = new JFileChooser();
-        file.setDialogTitle("Load Character");
-        FileNameExtensionFilter filter = new FileNameExtensionFilter(".json", "json");
-        file.setFileFilter(filter);
-        int result = file.showOpenDialog(null);
-        if (result == JFileChooser.APPROVE_OPTION){
-            File selectedFile = file.getSelectedFile();
-            dataCall.loadData(selectedFile);
-            MenuManager.loadChar();
-        }
     }//GEN-LAST:event_jButton1ActionPerformed
 
     /**
@@ -133,27 +118,21 @@ public class WarningScreenLoad extends javax.swing.JFrame {
                 }
             }
         } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(WarningScreenLoad.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(WarningScreenNew.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(WarningScreenLoad.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(WarningScreenNew.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(WarningScreenLoad.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(WarningScreenNew.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(WarningScreenLoad.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(WarningScreenNew.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
         //</editor-fold>
         //</editor-fold>
 
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             public void run() {
-                new WarningScreenLoad().setVisible(true);
+                new WarningScreenNew().setVisible(true);
             }
         });
     }
